Possible temporary input method.
I changed the input method to try and fix an issue with the model returning the same index of the first index. This code doesn't align yet. Only error so far is a bad link of vocab, it used to input proper, please use older version until new version patches errors