Updated A lot(Read Description))
So I ran into some errors with the start and end tokens, the way it was done before stripped the <>. So this is partly to fix that, In that time I found this error by testing chat functionality. I managed to find the actual versions of the packages I'm using and so I updated those to try and make it work on my laptop, but it just doesn't work yet. Testing now, still functional to the last point