Minor fixes for generating responses
I tried to fix the response generation and it seems to be broken on both sides because of an error saying Data Cardinality is Ambiguous. Traced the error to the target_seq in both the Beam Search and the generate_response functions. This is giving (2, ) when it should be giving a dynamic output like the encoder and decoder are. Until these 3 align, no responses